requirements.txt
numpy
pandas
scikit-learn
python-Levenshtein
tensorflow
tensorflow-hub
seaborn
matplotlib
Flask
flask-ngrok
pyngrok
pytest